update uie readme
