+++ model_zoo/uie/deploy/python/requirements_gpu.txt
@@ -1,3 +1,3 @@
 onnx
 onnxconverter_common
-onnxruntime-gpu
+onnxruntime-gpu==1.12
